refactor(admin): extract MiniCalendar component
